#715 - move regex atom from eo.eolang to eo.eolang.txt package
