+++ eo-runtime/src/main/eo/org/eolang/string.eo
@@ -54,16 +54,3 @@
   # Converts this to hash
   [] > as-hash /int
 
-  # Converts this to regex.
-  # This string construction must starts and ends from "\" characters,
-  # then in parenthesis should be patter string and then, after the last
-  # "\" can be specified flag option (like this: "/(your-pattern)/s")
-  # e.g. ("/(word)/i".as-regex).matches "WORD" return TRUE
-  # other flags:
-  # /d - Enables Unix lines mode.
-  # /i - Enables case-insensitive matching.
-  # /x - Permits whitespace and comments in pattern.
-  # /m - Enables multiline mode.
-  # /s - Enables dotall mode.
-  # /u - Enables Unicode-aware case folding.
-  [] > as-regex /regex
